more differenciation between local and remote skype chat, because of serialization issues
