+++ src/main/java/com/skype/ChatUtils.java
@@ -2,10 +2,10 @@ package com.skype;
 
 /**
  *
- * @author Hans-Joachim Kliemeck <git@kliemeck.de>
+ * @author Hans-Joachim Kliemeck <klk@mmw.ag>
  */
 public class ChatUtils {
-	public static String getChatTopic(Chat chat) throws SkypeException {
-		return Utils.getProperty("CHAT", chat.getId(), "TOPIC");
-	}
+    public static String getChatTopic(Chat chat) throws SkypeException {
+        return Utils.getProperty("CHAT", chat.getId(), "TOPIC");
+    }
 }
