+++ src/main/java/hudson/plugins/skype/im/transport/SkypeChat.java
@@ -1,34 +1,13 @@
 package hudson.plugins.skype.im.transport;
 
-import com.skype.Chat;
-import com.skype.Chat.Status;
-import com.skype.SkypeException;
 import hudson.plugins.im.IMChat;
-import hudson.plugins.im.IMException;
 import hudson.plugins.im.IMMessageListener;
 
 /**
- * 1-on-1 Jabber chat.
  * 
  * @author kutzi
  */
-public class SkypeChat implements IMChat {
-
-    private final Chat chat;
-    private SkypeMessageListenerAdapter messageListener;
-
-    public SkypeChat(Chat chat) {
-        this.chat = chat;
-
-    }
-
-    public void sendMessage(String msg) throws IMException {
-        try {
-            this.chat.send(msg);
-        } catch (SkypeException e) {
-            throw new IMException(e);
-        }
-    }
+public abstract class SkypeChat implements IMChat {
 
     public String getNickName(String sender) {
         return sender;
@@ -48,14 +27,6 @@ public class SkypeChat implements IMChat {
 
         //SkypeImpl.removeChatMessageListener(messageListener);
 
-    }
-
-    public boolean isMultiUserChat() {
-        try {
-            return chat.getStatus().equals(Status.MULTI_SUBSCRIBED);
-        } catch (SkypeException ex) {
-            throw new RuntimeException(ex);
-        }
     }
 
     public String getIMId(String user) {
