+++ src/main/java/hudson/plugins/skype/im/transport/callables/BotCommandCallable.java
@@ -2,27 +2,30 @@ package hudson.plugins.skype.im.transport.callables;
 
 import com.skype.Chat;
 import com.skype.ChatMessage;
+import com.skype.Skype;
 import com.skype.SkypeException;
 import hudson.plugins.im.IMException;
 import hudson.plugins.im.IMMessage;
 import hudson.plugins.im.bot.Bot;
+import hudson.plugins.skype.im.transport.RemoteSkypeChat;
 import hudson.plugins.skype.im.transport.SkypeChat;
 import hudson.plugins.skype.im.transport.SkypeIMException;
 import hudson.remoting.Callable;
-import hudson.remoting.Channel;
+import java.io.Serializable;
 
 /**
  *
  * @author jbh
  */
-public class BotCommandCallable implements Callable<Boolean, SkypeIMException> {
-    
-    Chat chat = null;    
-    String senderId = null;
-    String toId = null;
-    String content = null;
+public class BotCommandCallable implements Callable<Boolean, SkypeIMException>, Serializable {
+    private SkypeChat skypeChat = null;
+    private String senderId = null;
+    private String toId = null;
+    private String content = null;
+
     public BotCommandCallable(Chat chat, ChatMessage msg) throws SkypeIMException {
-        this.chat = chat;
+        skypeChat = new RemoteSkypeChat(chat, msg);
+
         try {
             senderId = msg.getSenderId();
             toId = msg.getId();
@@ -33,29 +36,12 @@ public class BotCommandCallable implements Callable<Boolean, SkypeIMException> {
     }
 
     public Boolean call() throws SkypeIMException {
-        SkypeChat skypeChat = new SkypeChat(chat) {
-
-            @Override
-            public void sendMessage(String msg) throws IMException {
-                try {
-                    SkypeChatCallable sender = new SkypeChatCallable(new String[]{senderId}, msg);
-                    Channel.current().call(sender);                    
-                    
-                } catch (Exception ex) {
-                    throw new IMException(ex);
-                }
-            }            
-        };
-        Bot bot = new Bot(skypeChat, "hudson",
-                "hostname", "!", null);
-
-
+        Bot bot = new Bot(skypeChat, "hudson", "hostname", "!", null);
         if (content != null) {
             // replay original message:
-
             bot.onMessage(new IMMessage(senderId, toId, content));//Ask skype
-
         }
+
         return Boolean.TRUE;
     }
 }
